Updated automation compiler
- Now compiling automation scripts in bundles using glob, and
event-stream with gulp. This will allow the ability to add more
automation scripts in the future.
- Updated auto-join to search for element to be inserted with jQuery
rather than using a setInterval for memory usage reasons.
