+++ gulpfile.js
@@ -10,6 +10,7 @@ gulp       = require('gulp');
 del        = require('del');
 jade       = require('gulp-jade');
 sass       = require('gulp-sass');
+rename     = require('gulp-rename');
 concat     = require('concat');
 neat       = require('node-neat');
 browserify = require('browserify');
@@ -18,6 +19,8 @@ source     = require('vinyl-source-stream');
 strictify  = require('strictify');
 babelify   = require('babelify');
 opn        = require('opn');
+es         = require('event-stream');
+glob       = require('glob');
 
 //Require our tasks since we have required our global plugins:
 var requireDir = require('require-dir');
