+++ package.json
@@ -9,10 +9,13 @@
     "browserify": "^11.0.0",
     "concat": "^1.0.0",
     "del": "^1.2.0",
+    "event-stream": "^3.3.2",
     "font-awesome": "^4.3.0",
+    "glob": "^6.0.4",
     "gulp": "^3.9.0",
     "gulp-jade": "^1.0.1",
     "gulp-load-plugins": "^1.2.0",
+    "gulp-rename": "^1.2.2",
     "gulp-sass": "^2.0.4",
     "lodash": "^4.2.1",
     "material-ui": "^0.14.4",
