--- src/js/automation/automation.js
@@ -1,5 +1,0 @@
-import chromeStorage from '../lib/chrome-storage';
-import autoJoin from './classes/autoJoin.js';
-
-const AutoJoin = new autoJoin();
-AutoJoin.initialize();
