+++ tasks/compile-automation.js
@@ -1,12 +1,22 @@
-gulp.task('compile:automation', function() {
-  var b = browserify({ entries: paths.automation }, {
-    debug: true
-  }).transform(babelify, {
-    presets: ["es2015", "react"]
-  });
+gulp.task('compile:automation', function(done) {
+  glob(paths.automation, function(err, files) {
+    if(err) done(err);
 
-  return b.bundle()
-    .pipe(source('automation.js'))
-    .pipe(buffer())
-    .pipe(gulp.dest(paths.output.automation));
+    var tasks = files.map(function(entry) {
+      return browserify({ entries: [entry] })
+        .transform(babelify, {
+          presets: ["es2015", "react"]
+        })
+        .bundle()
+        .pipe(source(entry))
+        .pipe(buffer())
+        .pipe(rename(function(path) {
+          path.extname = '.bundle.js';
+          path.dirname = '';
+          path.basename = path.basename.replace('script-', '');
+        }))
+        .pipe(gulp.dest(paths.output.automation));
+      });
+    es.merge(tasks).on('end', done);
+  })
 });
